arm64: dts: qcom: sm8150: Delete mem-offline node to fully disable PASR
Although the `mem_offline.nopasr=1` module parameter can be used to disable
the PASR memory-offline feature, the bootloader still passes a `mem=`
parameter on the kernel command line to restrict the total amount of RAM
available.

With 12 GiB of RAM, the bootloader appends `mem=8139008768B` to the kernel
command line, thereby limiting system RAM to 8 GiB. This limitation is
never lifted when PASR is disabled via `mem_offline.nopasr=1`, resulting in
the device being limited to 6.8 GiB of RAM indefinitely.

The bootloader checks for the presence of the mem-offline node in the
device tree to determine whether or not it appends command line arguments
related to PASR.

Therefore, in order to fully disable PASR without losing about half of the
system's total RAM, delete the mem-offline node from the device tree.

Signed-off-by: Sultan Alsawaf <sultan@osomprivacy.com>
Change-Id: I47a2141dfdc3ac900ef9d04151ad2ff661e0bdaf